error handling

+++ dbtest.cc
@@ -32,16 +32,21 @@ main()
 
     db.createNewsgroup("A");
 
+    if(db.deleteNewsgroup(1000)) {
+        cout << "Delete nonexistant should not be possible" << endl;
+        return 1;
+    }
+
     newsgroup = *db.getNewsgroups().begin();
     db.createArticle(newsgroup.ident, "TEST", "ME", "This is a test");
     auto arts = db.getArticles(newsgroup.ident);
     auto art = *arts.begin();
+    art = *db.getArticle(newsgroup.ident, art.ident);
     cout << "Should be TEST: ";
     cout << art.title << endl;
     if (art.title != "TEST") {
         return 1;
     }
-
 
     db.deleteArticle(newsgroup.ident, art.ident);
     arts = db.getArticles(newsgroup.ident);
